+++ cmd/gitlab-pages-deployer/transfer.go
@@ -13,21 +13,21 @@ func panicOnFileSystemError(err error) {
 	}
 }
 
-func renameNamespaceJob(message *workers.Msg, projectID int64, fullPathWas, fullPath string) {
+func renameNamespaceJob(message *workers.Msg, projectID int, fullPathWas, fullPath string) {
 	err := os.Rename(filepath.Join(*deployerRoot, fullPathWas), filepath.Join(*deployerRoot, fullPath))
 	panicOnFileSystemError(err)
 
 	touchDaemon()
 }
 
-func renameProjectJob(message *workers.Msg, projectID int64, pathWas, path, fullPath string) {
+func renameProjectJob(message *workers.Msg, projectID int, pathWas, path, fullPath string) {
 	err := os.Rename(filepath.Join(*deployerRoot, fullPath, pathWas), filepath.Join(*deployerRoot, fullPath, path))
 	panicOnFileSystemError(err)
 
 	touchDaemon()
 }
 
-func moveProjectJob(message *workers.Msg, projectID int64, path, fullPathWas, fullPath string) {
+func moveProjectJob(message *workers.Msg, projectID int, path, fullPathWas, fullPath string) {
 	err := os.MkdirAll(filepath.Join(*deployerRoot, fullPath), 0750)
 	panicOnFileSystemError(err)
 
@@ -37,7 +37,7 @@ func moveProjectJob(message *workers.Msg, projectID int64, path, fullPathWas, fu
 	touchDaemon()
 }
 
-func removeJob(message *workers.Msg, projectID int64, namespacePath, path string) {
+func removeJob(message *workers.Msg, projectID int, namespacePath, path string) {
 	err := os.RemoveAll(filepath.Join(*deployerRoot, namespacePath, path))
 	panicOnFileSystemError(err)
